Overhauled driveBalance
The PID command structure sucks for balancing with a pitch angle. The PIDCommandControl will calculate the fastest route to the desired setpoint. When it comes to being balanced the fastest route is to backup and level out on the ground. Manual PID calculations are going to be used now to avoid the smart a$$ robot class.